adding pyinstaller stuff to SCM runs
